frontend/pyproject.toml
[project]
name = "frontend"
version = "0.1.0"
description = ""
authors = [
    {name = "PatStLouis",email = "patrick.st-louis@opsecid.ca"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
